G_ftell() and G_fseek(): off_t replaced by int (merge https://trac.osgeo.org/grass/changeset/45468 from devbr6)
git-svn-id: https://svn.osgeo.org/grass/grass/branches/releasebranch_6_4@45935 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ include/gisdefs.h
@@ -1083,8 +1083,8 @@ void G_rotate_around_point(double, double, double *, double *, double);
 void G_rotate_around_point_int(int, int, int *, int *, double);
 
 /* seek.c */
-off_t G_ftell(FILE *);
-void G_fseek(FILE *, off_t, int);
+int G_ftell(FILE *);
+void G_fseek(FILE *, int, int);
 
 /* sample.c */
 DCELL G_get_raster_sample_nearest(
